[Feat] 관리자 대시보드, 상세 통계 하드코딩 값 제거 및 기능 구현

+++ src/main/java/org/example/backend/foodpick/domain/user/dto/PieItem.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class PieItem {
     private String name;
-    private int population;
+    private Long population;
 }
